Time: 98 ms (63.59%), Space: 6 MB (22.83%) - LeetHub

+++ 1680-concatenation-of-consecutive-binary-numbers/1680-concatenation-of-consecutive-binary-numbers.cpp
@@ -4,7 +4,7 @@ public:
         return log2(n)+1;
     }
     int concatenatedBinary(int n) {
-        long long res = 0;
+        long res = 0;
         int m = 1e9+7;
         for(int i=1; i<=n; i++){
             int len = nofbits(i);
